removed directory API route, moved SQL to getStaticProps
